feat: add site crawler, dashboard redesign, SEO tooling
- Site crawler: sitemap parser, HTTP fetcher, 6 analyzers (http, onpage,
  images, links, content, resources), engine with post-processing,
  report generation, diff between runs, GSC prioritization
- Dashboard: 9 pages (overview, SEO audit with pillar score, live canvas,
  site crawler with progress bar, piano editoriale with 3-tab UX,
  comandi pipeline, ricerca semantica)
- SEO: pillar score system, MedicalCondition schema.org for pillars,
  Yoast title/meta description management via REST API
- External link checker with browser headers and false positive filtering
- Removed TUI (replaced by Streamlit dashboard)
- Simplified CLI: mnemosyne crawl, mnemosyne dashboard
- WP plugin: mnemosyne-meta for Yoast REST API access
- 113 tests passing

+++ mnemosyne/embeddings/generator.py
@@ -35,7 +35,7 @@ def generate_embeddings(
 
         print(f"Generating embedding for post {post_id}...")
         response = openai_client.embeddings.create(input=text, model=model)
-        vector = np.array(response.data[0].embedding, dtype=np.float64)
+        vector = np.array(response.data[0].embedding, dtype=np.float32)
 
         conn.execute(
             "INSERT OR REPLACE INTO embeddings (post_id, model_name, vector, source_hash, created_at) "
